ecommerce_api/requirements.txt
Django==5.1.1
django-axes==7.0.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
mysqlclient==2.2.4
PyMySQL==1.1.1
sqlparse==0.5.1
cryptography==43.0.1
psycopg2-binary==2.9.3
django-environ==0.8.1